// ui/src/generated/types/deviceResponse.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */
import type { Device } from './device'

export interface DeviceResponse {
  description: string
  device: Device
  success: boolean
}