// ui/src/generated/types/getUsersParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */

export type GetUsersParams = {
  /**
   * Comma-separated list of user fields; only listed fields will be returned in response
   */
  fields?: string
}