// ui/src/generated/types/getGroupsParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */

export type GetGroupsParams = {
  /**
   * Comma-seperated list of fields; only listed fields will be returned in response
   */
  fields?: string
  /**
   * Selects the groups for which you are the owner (true) or a simple member (false); note that by not providing this parameter, it means all groups to which you belong are selected
   */
  owner?: boolean
}